factor move categorisation calculation out of AI Review component
